refactor: move all tests to a single file
